integrate templateiterator and logginghook into edflow
