+++ examples/template_tfe/edflow.py
@@ -1,139 +1,9 @@
 import functools
-from edflow.iterators.model_iterator import PyHookedModelIterator
-from edflow.hooks.checkpoint_hooks.lambda_checkpoint_hook import LambdaCheckpointHook
-from edflow.eval.pipeline import EvalHook
-from edflow.project_manager import ProjectManager
 import tensorflow as tf
+tf.enable_eager_execution()
 import tensorflow.keras as tfk
 import numpy as np
-
-tf.enable_eager_execution()
-
-
-# ================ Generic Logging Hook ==================
-
-
-from edflow.hooks.hook import Hook
-from edflow.util import retrieve
-from edflow.custom_logging import get_logger
-from edflow.iterators.batches import plot_batch
-import os
-class LoggingHook(Hook):
-    """Evaluate logging ops at an interval of training steps."""
-    def __init__(
-        self,
-        paths,
-        interval,
-        root_path,
-    ):
-        """Args:
-            paths (list(str)): List of key-paths to logging outputs. Will be
-                expanded so they can be evaluated lazily.
-            interval (int): Intervall of training steps before logging.
-            root_path (str): Path at which the logs are stored.
-        """
-        self.paths = paths
-        self.interval = interval
-        self.root = root_path
-        self.logger = get_logger(self)
-        self.handlers = {
-                "images": self.log_images,
-                "scalars": self.log_scalars}
-
-    def after_step(self, batch_index, last_results):
-        if batch_index % self.interval == 0:
-            self._step = last_results["global_step"]
-            self.logger.info("global_step: {}".format(self._step))
-            for path in self.paths:
-                for k in self.handlers:
-                    handler_results = retrieve(last_results, path+"/"+k, default = dict())
-                    self.handlers[k](handler_results)
-            self.logger.info("project root: {}".format(self.root))
-
-    def log_scalars(self, results):
-        for name in sorted(results.keys()):
-            self.logger.info("{}: {}".format(name, results[name]))
-
-    def log_images(self, results):
-        for name, image_batch in results.items():
-            full_name = name + "_{:07}.png".format(self._step)
-            save_path = os.path.join(self.root, full_name)
-            plot_batch(image_batch, save_path)
-
-
-# ================ Generic Iterator ==================
-
-
-from edflow.hooks.util_hooks import IntervalHook
-
-class TemplateIterator(PyHookedModelIterator):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        # and wrap it into a LambdaCheckpointHook
-        self.ckpthook = LambdaCheckpointHook(
-            root_path=ProjectManager.checkpoints,
-            global_step_getter=self.get_global_step,
-            global_step_setter=self.set_global_step,
-            save=self.save,
-            restore=self.restore,
-            interval=self.config.get("ckpt_freq", None),
-        )
-        if not self.config.get("test_mode", False):
-            self._train_ops = self.config.get("train_ops", ["step_ops/train_op"])
-            self._log_ops = self.config.get("log_ops", ["step_ops/log_op"])
-            # logging
-            self.loghook = LoggingHook(
-                    paths=self._log_ops,
-                    root_path=ProjectManager.train,
-                    interval=1,
-                    )
-            # wrap it in interval hook
-            self.ihook = IntervalHook(
-                [self.loghook],
-                interval=self.config.get("start_log_freq", 1),
-                modify_each=1,
-                max_interval=self.config.get("log_freq", 1000),
-                get_step=self.get_global_step,
-            )
-            self.hooks.append(self.ihook)
-            # write checkpoints after epoch or when interrupted
-            self.hooks.append(self.ckpthook)
-        else:
-            # evaluate
-            self._eval_op = self.config.get("eval_op", "step_ops/eval_op")
-            self.evalhook = EvalHook(
-                    dataset = self.dataset,
-                    step_getter = self.get_global_step,
-                    keypath = self._eval_op,
-                    meta = self.config)
-            self.hooks.append(self.evalhook)
-            self._train_ops = []
-            self._log_ops = []
-
-    def initialize(self, checkpoint_path=None):
-        if checkpoint_path is not None:
-            self.ckpthook(checkpoint_path)
-
-    def step_ops(self):
-        return self.step_op
-
-    def run(self, fetches, feed_dict):
-        results = super().run(fetches, feed_dict)
-        for train_op in self._train_ops:
-            retrieve(results, train_op)()
-        return results
-
-    def save(self, checkpoint_path):
-        raise NotImplemented()
-
-    def restore(self, checkpoint_path):
-        raise NotImplemented()
-
-    def step_op(self, model, **kwargs):
-        raise NotImplemented()
-
-
-# ================ Application Code ==================
+from edflow import TemplateIterator, get_logger
 
 
 class Model(tfk.Model):
